fix : sku weight mongo update
