+++ jeecg-module-system/jeecg-system-biz/src/main/java/org/jeecg/modules/business/mongoRepository/SkuRepository.java
@@ -3,6 +3,7 @@ package org.jeecg.modules.business.mongoRepository;
 import org.jeecg.modules.business.model.SkuDocument;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.data.mongodb.repository.Query;
+import java.util.List;
 
 /**
  * @Description: Mongo Repository for Sku Collection
@@ -11,10 +12,10 @@ import org.springframework.data.mongodb.repository.Query;
  */
 public interface SkuRepository extends MongoRepository<SkuDocument, String> {
     @Query("{'erpCode': ?0}")
-    SkuDocument findByErpCode(String erpCode);
+    List<SkuDocument> findByErpCode(String erpCode);
 
     @Query("{'skuId': ?0}")
-    SkuDocument findBySkuId(String skuId);
+    List<SkuDocument> findBySkuId(String skuId);
 
     long count();
